refactor translator code location
