+++ src/aa_pbs_exporter/pbs_2022_01/helpers.py
@@ -5,7 +5,10 @@ from typing import Callable, Iterable, Sequence
 
 from aa_pbs_exporter.pbs_2022_01.models import raw
 from aa_pbs_exporter.pbs_2022_01.models.common import HashedFile
-from aa_pbs_exporter.pbs_2022_01.result_handler import DebugToFile, ParsedToRaw
+from aa_pbs_exporter.pbs_2022_01.translate.result_handler import (
+    DebugToFile,
+    ParsedToRaw,
+)
 from aa_pbs_exporter.pbs_2022_01.validate_raw import RawValidator
 from aa_pbs_exporter.snippets.file.validate_file_out import validate_file_out
 from aa_pbs_exporter.snippets.hash.file_hash import make_hashed_file
