+++ tests/aa_pbs_exporter/pbs_2022_01/translate/to_compact/one_page/one_page_to_compact_test.py
@@ -9,7 +9,7 @@ from aa_pbs_exporter.airports.airports import tz_name_from_iata
 from aa_pbs_exporter.pbs_2022_01.helpers import debug_parse_raw_bidpackage
 from aa_pbs_exporter.pbs_2022_01.models import compact, raw
 from aa_pbs_exporter.pbs_2022_01.parse_manager import ParseManager
-from aa_pbs_exporter.pbs_2022_01.raw_to_compact import RawToCompact
+from aa_pbs_exporter.pbs_2022_01.translate.raw_to_compact import RawToCompact
 from aa_pbs_exporter.pbs_2022_01.validate_compact import CompactValidator
 from aa_pbs_exporter.snippets.file.validate_file_out import validate_file_out
